refactor: rename module compute_dti_reference_values to compute_dti_normative_values
